+++ ksef-client/src/main/java/pl/akmf/ksef/sdk/client/Headers.java
@@ -16,5 +16,6 @@ public final class Headers {
     public static final String APPLICATION_XML =  "application/xml";
     public static final String BEARER = "Bearer ";
     public static final String X_KSEF_FEATURE = "X-KSeF-Feature";
+    public static final String ENFORCE_XADES_COMPLIANCE = "enforce-xades-compliance";
     public static final String RETRY_AFTER = "Retry-After";
 }